digit_classifier/pyproject.toml
[project]
name = "jersey-digit-classifier"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "torch",
    "torchvision",
    "pytorch-lightning",
    "movinets @ https://github.com/Atze00/MoViNet-pytorch/archive/refs/heads/main.zip",
    "wandb",
    "python-dotenv",
    "scikit-learn",
    "Pillow",
    "numpy",
]

[tool.uv]
package = false
